% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clga.R
\docType{package}
\name{clga}
\alias{clga}
\title{clga: Download Traffic Data for Commons Library content from Google Analytics}
\description{
This package contains a suite of functions for downloading data on traffic
to research briefing landing pages and to the Commons Library website for
research products produced by the House of Commons Library.
}
