% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic.R
\name{fetch_briefing_traffic}
\alias{fetch_briefing_traffic}
\title{Download traffic data for a research briefing with the given URL}
\usage{
fetch_briefing_traffic(url, start_date, end_date)
}
\arguments{
\item{url}{The URL of a page for which traffic data is requested.}

\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}
}
\value{
A tibble of users, sessions, and pageviews by date.
}
\description{
\code{fetch_briefing_traffic} downloads data showing the number of users,
sessions, and pageviews for a given research briefing url during the given
dates and returns the data as a tibble.
}
