% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic.R
\name{fetch_briefings_traffic}
\alias{fetch_briefings_traffic}
\title{Download traffic data for all pages in the research briefings view}
\usage{
fetch_briefings_traffic(start_date, end_date, detailed = FALSE,
  dim_filters = NULL)
}
\arguments{
\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}

\item{detailed}{A boolean indicating whether to return the results by date
in total, or broken down by individual page. The default is FALSE.}

\item{dim_filters}{A set of dimension filters to constrain the results. The
default is NULL.}
}
\value{
A tibble of traffic metrics.
}
\description{
\code{fetch_briefings_traffic} downloads data on traffic metrics for all
research briefings under the research briefings view during the given dates
and returns the data as a tibble.
}
\details{
Note that this is not all research briefings traffic, as it does not include
traffic to the research briefings pages on the Parliamentary intranet.
}
