% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic.R
\name{fetch_cbp_traffic_intranet}
\alias{fetch_cbp_traffic_intranet}
\title{Download traffic data for all Commons briefing paper landing pages on
the Parliamentary intranet}
\usage{
fetch_cbp_traffic_intranet(start_date, end_date, detailed = FALSE)
}
\arguments{
\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}

\item{detailed}{A boolean indicating whether to return the results by date
in total, or broken down by date and page. The default is FALSE.}

\item{internal}{A boolean indicating whether to return only the results for
traffic from internal parliamentary networks. The default is FALSE.}
}
\value{
A tibble of traffic metrics.
}
\description{
\code{fetch_cbp_traffic_intranet} downloads data on traffic metrics for all
Commons Library briefing papers on the intranet during the given dates and
returns the data as a tibble. These are the landing pages for briefings
whose codes begin with "SN" or "CBP".
}
\details{
Note that this is not all research briefings traffic, as it does not include
traffic to the research briefings pages on the main Parliament website. Use
\code{fetch_cbp_traffic_public} to retrieve equivalent data for the main
Parliament website.
}
