% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons.R
\name{fetch_cdp_traffic_intranet}
\alias{fetch_cdp_traffic_intranet}
\title{Download traffic data for all Commons debate pack landing pages on the
Parliamentary intranet}
\usage{
fetch_cdp_traffic_intranet(start_date, end_date, detailed = FALSE)
}
\arguments{
\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}

\item{detailed}{A boolean indicating whether to return the results by date
in total, or broken down by date and page. The default is FALSE.}
}
\value{
A tibble of traffic metrics.
}
\description{
\code{fetch_cdp_traffic_intranet} downloads data on traffic metrics for all
Commons Library debate packs on the intranet during the given dates and
returns the data as a tibble. These are the landing pages for briefings
whose ids begin with "CDP".
}
\details{
Note that this is not all debate pack traffic, as it does not include
traffic to the debate pack pages on the main Parliament website. Use
\code{fetch_cdp_traffic_public} to retrieve equivalent data for the main
Parliament website.
}
