% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons-downloads.R
\name{fetch_crb_downloads}
\alias{fetch_crb_downloads}
\title{Download data on document downloads for all Commons research briefings in
both the research briefings view and the research briefings intranet view}
\usage{
fetch_crb_downloads(
  start_date,
  end_date,
  internal = FALSE,
  by_date = FALSE,
  by_page = FALSE,
  combine = FALSE,
  merge_paths = FALSE,
  anti_sample = FALSE,
  use_resource_quotas = FALSE
)
}
\arguments{
\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}

\item{internal}{A boolean indicating whether to return only the results for
downloads from internal parliamentary networks. The default is FALSE.}

\item{by_date}{A boolean indicating whether to return the results broken
down by date. The default is FALSE.}

\item{by_page}{A boolean indicating whether to return the results broken
down by page. The default is FALSE.}

\item{combine}{A boolean indicating whether to combine the totals from
different properties or to report them separately. The default is FALSE.}

\item{merge_paths}{A boolean indicating whether to aggregate figures for all
pages that have the same root path i.e. for all pages whose paths differ
only by their query strings or internal anchors. This parameter is ignored
if \code{by_page} is set to FALSE. The default value is FALSE.}

\item{anti_sample}{A boolean indicating whether to use googleAnalyticsR's
anti-sample feature, which chunks API calls to keep the number of records
requested under the API limits that trigger sampling. This makes the
download process slower but ensures that all records are returned. Only
use this feature if you see that an API request triggers sampling without
it. The default is FALSE.}

\item{use_resource_quotas}{A boolean indicating whether to use the resource
quotas in Parliament's Google Analytics account to prevent sampling.
This is a faster and more effective way to disable sampling than using
\code{anti_sample}, but using resource quotas consumes tokens from a
limited daily quota. Use this when \code{anti_sample} still fails to
prevent sampling or is taking too long. Note that using resource quotas
takes precendence over anti-samping: if \code{use_resource_quotas} is TRUE
\code{anti_sample} is automatically set to FALSE. The default is FALSE.}
}
\value{
A tibble of traffic metrics.
}
\description{
\code{fetch_crb_downloads} downloads data on document download metrics for
all Commons research briefings on both the public Parliament website and the
Parliamentary intranet during the given dates and returns the data as a
tibble.
}
\details{
The data can either be combined so that each result appears once with
totals across all properties for the Parliament website and the intranet,
or reported separately with separate rows for each property.

By default, download figures are reported separately for each website
property in Google Analytics that contains some of the requested data. You
can use the \code{combine} argument to optionally combine download figures
so that each result appears only once with figures totalled across all
relevant properties.

Download figures can be requested by page. Google Analytics treats webpages
requested with different query strings and section anchors as different
pages in its downloads data. You can use the \code{merge_paths} argument to
optionally sum the figures for pages with the same base path.

If \code{combine} and \code{merge_paths} are both set to TRUE, rows for
different properties are combined before paths are merged.
}
