% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons-downloads.R
\name{fetch_crb_downloads_intranet}
\alias{fetch_crb_downloads_intranet}
\title{Download data on document downloads for all Commons research briefing
landing pages on the Parliamentary intranet}
\usage{
fetch_crb_downloads_intranet(
  start_date,
  end_date,
  by_date = FALSE,
  by_page = FALSE,
  merge_paths = FALSE,
  anti_sample = FALSE,
  use_resource_quotas = FALSE
)
}
\arguments{
\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}

\item{by_date}{A boolean indicating whether to return the results broken
down by date. The default is FALSE.}

\item{by_page}{A boolean indicating whether to return the results broken
down by page. The default is FALSE.}

\item{merge_paths}{A boolean indicating whether to aggregate figures for all
pages that have the same root path i.e. for all pages whose paths differ
only by their query strings or internal anchors. This parameter is ignored
if \code{by_page} is set to FALSE. The default value is FALSE.}

\item{anti_sample}{A boolean indicating whether to use googleAnalyticsR's
anti-sample feature, which chunks API calls to keep the number of records
requested under the API limits that trigger sampling. This makes the
download process slower but ensures that all records are returned. Only
use this feature if you see that an API request triggers sampling without
it. The default is FALSE.}

\item{use_resource_quotas}{A boolean indicating whether to use the resource
quotas in Parliament's Google Analytics account to prevent sampling.
This is a faster and more effective way to disable sampling than using
\code{anti_sample}, but using resource quotas consumes tokens from a
limited daily quota. Use this when \code{anti_sample} still fails to
prevent sampling or is taking too long. Note that using resource quotas
takes precendence over anti-samping: if \code{use_resource_quotas} is TRUE
\code{anti_sample} is automatically set to FALSE. The default is FALSE.}
}
\value{
A tibble of document download metrics.
}
\description{
\code{fetch_crb_downloads_intranet} downloads data on document download
metrics for all Commons Library research briefings on the intranet during
the given dates and returns the data as a tibble. These are the landing
pages for briefings whose ids begin with "SN", "CBP", or "CDP".
}
\details{
Note that this is not all research briefings downloads, as it does not
include downloads on the research briefings pages on the main Parliament
website. Use \code{fetch_crb_downloads_public} to retrieve equivalent data
for the main Parliament website.

Download figures can be requested by page. Google Analytics treats webpages
requested with different query strings and section anchors as different
pages in its downloads data. You can use the \code{merge_paths} argument to
optionally sum the figures for pages with the same base path.
}
