% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons.R
\name{fetch_crb_traffic}
\alias{fetch_crb_traffic}
\title{Download traffic data for all Commons research briefings in both the
research briefings view and the research briefings intranet view}
\usage{
fetch_crb_traffic(start_date, end_date, internal = FALSE,
  detailed = FALSE, combined = TRUE)
}
\arguments{
\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}

\item{internal}{A boolean indicating whether to return only the results for
traffic from internal parliamentary networks. The default is FALSE.}

\item{detailed}{A boolean indicating whether to return the results by date
in total, or broken down by individual page. The default is FALSE.}

\item{combined}{A boolean indicating whether to combine the totals from
the website and the intranet or to report them separately. The default is
TRUE.}
}
\value{
A tibble of traffic metrics.
}
\description{
\code{fetch_crb_traffic} downloads data on traffic metrics for all Commons
research briefings on both the public Parliament website and the
Parliamentary intranet during the given dates and returns the data as a
tibble.
}
\details{
The data can either be combined so that each result appears once with
totals across both the Parliament website and the intranet, or reported
separately with separate rows for the website and the intranet.
}
