% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/briefings-events.R
\name{fetch_downloads_for_rb}
\alias{fetch_downloads_for_rb}
\title{Download data on document downloads for a research briefing with the given
URL on the main Parliament website and the Parliamentary intranet}
\usage{
fetch_downloads_for_rb(
  url,
  start_date,
  end_date,
  internal = FALSE,
  by_date = FALSE,
  combine = FALSE,
  anti_sample = FALSE,
  use_resource_quotas = FALSE
)
}
\arguments{
\item{url}{The URL of a page for which downloads data is requested.}

\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}

\item{internal}{A boolean indicating whether to return only the results for
downloads from internal parliamentary networks. The default is FALSE.}

\item{by_date}{A boolean indicating whether to return the results broken
down by date. The default is FALSE.}

\item{combine}{A boolean indicating whether to combine the totals from
different properties or to report them separately. The default is FALSE.}

\item{anti_sample}{A boolean indicating whether to use googleAnalyticsR's
anti-sample feature, which chunks API calls to keep the number of records
requested under the API limits that trigger sampling. This makes the
download process slower but ensures that all records are returned. Only
use this feature if you see that an API request triggers sampling without
it. The default is FALSE.}

\item{use_resource_quotas}{A boolean indicating whether to use the resource
quotas in Parliament's Google Analytics account to prevent sampling.
This is a faster and more effective way to disable sampling than using
\code{anti_sample}, but using resource quotas consumes tokens from a
limited daily quota. Use this when \code{anti_sample} still fails to
prevent sampling or is taking too long. Note that using resource quotas
takes precendence over anti-samping: if \code{use_resource_quotas} is TRUE
\code{anti_sample} is automatically set to FALSE. The default is FALSE.}
}
\value{
A tibble of document download metrics.
}
\description{
\code{fetch_downloads_for_rb} downloads data on document download metrics
for a given research briefing url on both the main Parliament website and
the Parliamentary intranet during the given dates and returns the data as a
tibble.
}
\details{
By default, download figures are reported separately for each website
property in Google Analytics that contains some of the requested data. You
can use the \code{combine} argument to optionally combine download figures
so that each result appears only once with figures totalled across all
relevant properties.
}
