% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{fetch_events}
\alias{fetch_events}
\title{Download data on document downloads for a given view}
\usage{
fetch_events(
  view_id,
  start_date,
  end_date,
  dimensions = NULL,
  dim_filters = NULL,
  anti_sample = FALSE,
  use_resource_quotas = FALSE
)
}
\arguments{
\item{view_id}{The view id in Google Analytics}

\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}

\item{dimensions}{A vector of dimensions to return. The default is NULL.}

\item{dim_filters}{A set of dimension filters to constrain the results. The
default is NULL.}

\item{anti_sample}{A boolean indicating whether to use googleAnalyticsR's
anti-sample feature, which chunks API calls to keep the number of records
requested under the API limits that trigger sampling. This makes the
download process slower but ensures that all records are returned. Only
use this feature if you see that an API request triggers sampling without
it. The default is FALSE.}

\item{use_resource_quotas}{A boolean indicating whether to use the resource
quotas in Parliament's Google Analytics account to prevent sampling.
This is a faster and more effective way to disable sampling than using
\code{anti_sample}, but using resource quotas consumes tokens from a
limited daily quota. Use this when \code{anti_sample} still fails to
prevent sampling or is taking too long. Note that using resource quotas
takes precendence over anti-samping: if \code{use_resource_quotas} is TRUE
\code{anti_sample} is automatically set to FALSE. The default is FALSE.}
}
\value{
A tibble of event metrics.
}
\description{
\code{fetch_events} downloads data on event metrics for a given
view during the given dates and returns the data as a tibble.
}
