% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsite.R
\name{fetch_ms_traffic_by_type}
\alias{fetch_ms_traffic_by_type}
\title{Download traffic data for different types of posts on the Commons Library
microsite}
\usage{
fetch_ms_traffic_by_type(type_regexp, start_date, end_date,
  dimensions = NULL, internal = FALSE, by_date = FALSE,
  by_page = FALSE, merge_paths = FALSE, anti_sample = FALSE)
}
\arguments{
\item{type_regexp}{A regular expression that describes the page path for one
or more post types.}

\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}

\item{dimensions}{A set of dimensions to constrain the results. The
default is NULL.}

\item{internal}{A boolean indicating whether to return only the results for
traffic from internal parliamentary networks. The default is FALSE.}

\item{by_date}{A boolean indicating whether to return the results broken
down by date. The default is FALSE.}

\item{by_page}{A boolean indicating whether to return the results broken
down by individual page. The default is FALSE.}

\item{merge_paths}{A boolean indicating whether to aggregate metrics for all
pages that have the same root path i.e. for all pages whose paths differ
only by their query string. This parameter is ignored if \code{by_page} is
set to FALSE. Note that while merging paths is necessary for analysis of
individual pages it can introduce small errors in the number of users by
page, as the same user may visit the same page through URLs with different
query strings. The default value is FALSE.}

\item{anti_sample}{A boolean indicating whether to use googleAnalyticsR's
anti-sample feature, which chunks API calls to keep the number of records
requested under the API limits that trigger sampling. This makes the
download process slower but ensures that all records are returned. Only
use this feature if you see that an API request triggers sampling without
it. The default is FALSE.}
}
\value{
A tibble of traffic metrics.
}
\description{
\code{fetch_ms_traffic_by_type} downloads data on traffic metrics for
specific types of posts based on their path prefixes during the given dates
and returns the data as a tibble. The specific types of posts to return are
defined with a regular expression that identifies their prefixes within the
page path.
}
