% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic.R
\name{fetch_page_traffic}
\alias{fetch_page_traffic}
\title{Download traffic data for a page with the given URL and view id}
\usage{
fetch_page_traffic(url, view_id, start_date, end_date)
}
\arguments{
\item{url}{The URL of a page.}

\item{view_id}{The view id in Google Analytics for the requested page.}

\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}
}
\value{
A tibble of users, sessions, and pageviews by date.
}
\description{
\code{fetch_page_traffic} downloads data showing the number of users,
sessions, and pageviews for a given url during the given dates and returns
the data as a tibble.
}
\keyword{internal}
