% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/briefings.R
\name{fetch_rb_traffic_all_sources}
\alias{fetch_rb_traffic_all_sources}
\title{Download traffic data for pages in both the research briefings view and the
research briefings intranet view using the given functions for each view}
\usage{
fetch_rb_traffic_all_sources(fetch_public, fetch_intranet, start_date,
  end_date, internal = FALSE, detailed = FALSE, combined = TRUE)
}
\arguments{
\item{fetch_public}{A \code{fetch_*_public} function for a set of briefings.}

\item{fetch_intranet}{A \code{fetch_*_intranet} function for a set of
briefings. For this function to behave as expected \code{fetch_public} and
\code{fetch_intranet} should target the same set of briefings in each view.}

\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}

\item{internal}{A boolean indicating whether to return only the results for
traffic from internal parliamentary networks. The default is FALSE.}

\item{detailed}{A boolean indicating whether to return the results by date
in total, or broken down by individual page. The default is FALSE.}

\item{combined}{A boolean indicating whether to combine the totals from
the website and the intranet or to report them separately. The default is
TRUE.}
}
\value{
A tibble of traffic metrics.
}
\description{
\code{fetch_rb_traffic_all_sources} downloads data on traffic metrics for
the same set of research briefings on both the public Parliament website and
the Parliamentary intranet during the given dates and returns the data as a
tibble. You must provide matching fetch_* functions for each view.
}
\details{
The data can either be combined so that each result appears once with
totals across both the Parliament website and the intranet, or reported
separately with separate rows for the website and the intranet.
}
\keyword{internal}
