% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic.R
\name{fetch_rb_traffic_intranet}
\alias{fetch_rb_traffic_intranet}
\title{Download traffic data for all pages in the research briefings intranet view}
\usage{
fetch_rb_traffic_intranet(start_date, end_date, detailed = FALSE,
  dim_filters = NULL)
}
\arguments{
\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}

\item{detailed}{A boolean indicating whether to return the results by date
in total, or broken down by individual page. The default is FALSE.}

\item{dim_filters}{A set of dimension filters to constrain the results. The
default is NULL.}
}
\value{
A tibble of traffic metrics.
}
\description{
\code{fetch_rb_traffic_intranet} downloads data on traffic metrics for all
research briefings on the Parliamentary intranet during the given dates and
returns the data as a tibble.
}
\details{
Note that this is not all research briefings traffic, as it does not include
traffic to the research briefings pages on the main Parliament website. Use
\code{fetch_rb_traffic_public} to retrieve equivalent data for the main
Parliament website.
}
