% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/briefings.R
\name{fetch_rb_traffic_public_by_filter}
\alias{fetch_rb_traffic_public_by_filter}
\title{Download traffic data for all pages in the research briefings view with the
given filters}
\usage{
fetch_rb_traffic_public_by_filter(start_date, end_date, detailed = FALSE,
  dim_filters = NULL)
}
\arguments{
\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}

\item{detailed}{A boolean indicating whether to return the results by date
in total, or broken down by individual page. The default is FALSE.}

\item{dim_filters}{A set of dimension filters to constrain the results. The
default is NULL.}
}
\value{
A tibble of traffic metrics.
}
\description{
\code{fetch_rb_traffic_public_by_filter} downloads data on traffic metrics
for all research briefings on the public Parliament website during the given
dates, with the given filters, and returns the data as a tibble.
}
\details{
Note that this is not all research briefings traffic, as it does not include
traffic to the research briefings pages on the Parliamentary intranet. Use
\code{fetch_rb_traffic_intranet_by_filter} to retrieve equivalent data for
the Parliamentary intranet.
}
