% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/briefings.R
\name{fetch_rb_traffic_public_by_type}
\alias{fetch_rb_traffic_public_by_type}
\title{Download traffic data for different types of Commons research briefings on
the main Parliamentary website}
\usage{
fetch_rb_traffic_public_by_type(type_regexp, start_date, end_date,
  internal = FALSE, detailed = FALSE, merge_paths = FALSE)
}
\arguments{
\item{type_regexp}{That regular expression that describes the page path for
one or more briefing types.}

\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}

\item{internal}{A boolean indicating whether to return only the results for
traffic from internal parliamentary networks. The default is FALSE.}

\item{detailed}{A boolean indicating whether to return the results by date
in total, or broken down by date and page. The default is FALSE.}

\item{merge_paths}{A boolean indicating whether to aggregate metrics for all
pages that have the same path i.e. for all pages whose paths differ only
by their query string. Note that while merging paths is necessary for
analysis of individual pages it can introduce small errors in the number
of users by page, as the same user may visit the same page through URLs
with different query strings. This parameter is ignored if \code{detailed}
is set to FALSE. The default value is FALSE.}
}
\value{
A tibble of traffic metrics.
}
\description{
\code{fetch_rb_traffic_public_by_type} downloads data on traffic metrics for
specific types of research breifing based on their id prefixes during the
given dates and returns the data as a tibble. The specific types of
briefings to return are defined with a regular expression that identifies
their id prefixes within the page path.
}
\details{
Note that this function only returns data for the main Parliament website.
Use \code{fetch_rb_traffic_intranet_by_type} to retrieve equivalent data
for the Parliamentary intranet.
}
