% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic.R
\name{fetch_traffic_for_page}
\alias{fetch_traffic_for_page}
\title{Download traffic data for a page with the given URL and view id}
\usage{
fetch_traffic_for_page(url, view_id, start_date, end_date)
}
\arguments{
\item{url}{The URL of a page.}

\item{view_id}{The view id in Google Analytics for the requested page.}

\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}
}
\value{
A tibble of traffic metrics.
}
\description{
\code{fetch_traffic_for_page} downloads data on traffic metrics for a given
url during the given dates and returns the data as a tibble.
}
\keyword{internal}
