% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic.R
\name{fetch_view_traffic}
\alias{fetch_view_traffic}
\title{Download traffic data for a given view}
\usage{
fetch_view_traffic(view_id, start_date, end_date)
}
\arguments{
\item{view_id}{The view id in Google Analytics}

\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}
}
\value{
A tibble of users, sessions, and pageviews by date.
}
\description{
\code{fetch_view_traffic} downloads data showing the number of users,
sessions, and pagevies for a given view during the given dates and returns
the data as a tibble.
}
\keyword{internal}
