% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_path_from_url}
\alias{get_path_from_url}
\title{Extract the path from a URL}
\usage{
get_path_from_url(url)
}
\arguments{
\item{url}{The url of a page.}
}
\value{
The path component of the URL.
}
\description{
\code{get_path_from_url} gets the path component from a URL. This is used to
get the unique id of a page within Google Analytics.
}
