% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label_events}
\alias{label_events}
\title{Label event metrics with the events they represent}
\usage{
label_events(events, event_name)
}
\arguments{
\item{events}{A tibble of event metrics returned from one of the event
fetching functions.}
}
\value{
A tibble whose data is identical to input tibble but whose event
  count columns are renamed with the given event name.
}
\description{
\code{label_events} takes a tibble of event metrics and re-labels the event
counts to represent the specific events in question.
}
\keyword{internal}
