% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_download_paths}
\alias{merge_download_paths}
\title{Aggregate document download metrics for all pages that have the same root
path}
\usage{
merge_download_paths(downloads, by_date)
}
\arguments{
\item{downloads}{A tibble of document download metrics by property, date and
page path returned from one of the download fetching functions.}

\item{by_date}{A boolean indicating whether the results are broken down by
date.}
}
\value{
A tibble with the same column structure as the input tibble where
  the metrics have been summed for all pages with the same root path.
}
\description{
\code{merge_download_paths} takes a tibble of document download metrics by
property, date and page and sums the metrics for all pages that have the
same path i.e. for all pages whose paths differ only by their query string
and internal links.
}
\keyword{internal}
