% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clga.R
\docType{package}
\name{clga}
\alias{clga}
\alias{clga-package}
\title{clga: Download Traffic Data for Commons Library content from Google Analytics}
\description{
clga: Download Traffic Data for Commons Library content from Google Analytics
}
