% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons.R
\name{fetch_cdp_traffic}
\alias{fetch_cdp_traffic}
\title{Download traffic data for all Commons debate packs in both the research
briefings view and the research briefings intranet view}
\usage{
fetch_cdp_traffic(start_date, end_date, internal = FALSE,
  by_date = FALSE, by_page = FALSE, merge_paths = FALSE,
  combined = FALSE, anti_sample = FALSE)
}
\arguments{
\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}

\item{internal}{A boolean indicating whether to return only the results for
traffic from internal parliamentary networks. The default is FALSE.}

\item{by_date}{A boolean indicating whether to return the results broken
down by date. The default is FALSE.}

\item{by_page}{A boolean indicating whether to return the results broken
down by page. The default is FALSE.}

\item{merge_paths}{A boolean indicating whether to aggregate metrics for all
pages that have the same root path i.e. for all pages whose paths differ
only by their query string. This parameter is ignored if \code{by_page} is
set to FALSE. Note that while merging paths is necessary for analysis of
individual pages it can introduce small errors in the number of users by
page, as the same user may visit the same page through URLs with different
query strings. The default value is FALSE.}

\item{combined}{A boolean indicating whether to combine the totals from
the website and the intranet or to report them separately. Note that
combining the traffic across both properties can introduce errors in
the number of users, as the same user may visit pages on both properties.
The default is FALSE.}

\item{anti_sample}{A boolean indicating whether to use googleAnalyticsR's
anti-sample feature, which chunks API calls to keep the number of records
requested under the API limits that trigger sampling. This makes the
download process slower but ensures that all records are returned. Only
use this feature if you see that an API request triggers sampling without
it. The default is FALSE.}
}
\value{
A tibble of traffic metrics.
}
\description{
\code{fetch_cdp_traffic} downloads data on traffic metrics for all Commons
debate packs on both the public Parliament website and the Parliamentary
intranet during the given dates and returns the data as a tibble.
}
\details{
The data can either be combined so that each result appears once with
totals across both the Parliament website and the intranet, or reported
separately with separate rows for the website and the intranet.
}
