% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons.R
\name{fetch_cdp_traffic_intranet}
\alias{fetch_cdp_traffic_intranet}
\title{Download traffic data for all Commons debate pack landing pages on the
Parliamentary intranet}
\usage{
fetch_cdp_traffic_intranet(start_date, end_date, by_date = FALSE,
  by_page = FALSE, merge_paths = FALSE, anti_sample = FALSE)
}
\arguments{
\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}

\item{by_date}{A boolean indicating whether to return the results broken
down by date. The default is FALSE.}

\item{by_page}{A boolean indicating whether to return the results broken
down by page. The default is FALSE.}

\item{merge_paths}{A boolean indicating whether to aggregate figures for all
pages that have the same root path i.e. for all pages whose paths differ
only by their query strings or internal anchors. This parameter is ignored
if \code{by_page} is set to FALSE. Note that while merging paths is
necessary for analysis of individual pages it can introduce errors in the
number of users by page, as the same user may visit the same page through
URLs with different query strings and anchors. The default value is FALSE.}

\item{anti_sample}{A boolean indicating whether to use googleAnalyticsR's
anti-sample feature, which chunks API calls to keep the number of records
requested under the API limits that trigger sampling. This makes the
download process slower but ensures that all records are returned. Only
use this feature if you see that an API request triggers sampling without
it. The default is FALSE.}
}
\value{
A tibble of traffic metrics.
}
\description{
\code{fetch_cdp_traffic_intranet} downloads data on traffic metrics for all
Commons Library debate packs on the intranet during the given dates and
returns the data as a tibble. These are the landing pages for briefings
whose ids begin with "CDP".
}
\details{
Note that this is not all debate pack traffic, as it does not include
traffic to the debate pack pages on the main Parliament website. Use
\code{fetch_cdp_traffic_public} to retrieve equivalent data for the main
Parliament website.

Traffic figures can be requested by page. Google Analytics treats webpages
requested with different query strings and section anchors as different
pages in its traffic data. You can use the \code{merge_paths} argument to
optionally sum the figures for pages with the same base path. This makes it
easier to calculcate the total number of views each distinct page has
received, but note that this may introduce errors in the number of users, as
a user may visit the same page using URLs with different query strings and
anchors.
}
