% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons.R
\name{fetch_crb_traffic_public}
\alias{fetch_crb_traffic_public}
\title{Download traffic data for all Commons research briefing landing pages on the
main Parliamentary website}
\usage{
fetch_crb_traffic_public(start_date, end_date, internal = FALSE,
  detailed = FALSE, merge_paths = FALSE)
}
\arguments{
\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}

\item{internal}{A boolean indicating whether to return only the results for
traffic from internal parliamentary networks. The default is FALSE.}

\item{detailed}{A boolean indicating whether to return the results by date
in total, or broken down by date and page. The default is FALSE.}

\item{merge_paths}{A boolean indicating whether to aggregate metrics for all
pages that have the same path i.e. for all pages whose paths differ only
by their query string. Note that while merging paths is necessary for
analysis of individual pages it can introduce small errors in the number
of users by page, as the same user may visit the same page through URLs
with different query strings. This parameter is ignored if \code{detailed}
is set to FALSE. The default value is FALSE.}
}
\value{
A tibble of traffic metrics.
}
\description{
\code{fetch_crb_traffic_public} downloads data on traffic metrics for all
Commons Library research briefings during the given dates and returns the
data as a tibble. These are the landing pages for briefings whose codes
begin with "SN", "CBP" or "CDP".
}
\details{
Note that this is not all Commons research briefing traffic, as it does not
include traffic to the research briefing pages on the Parliamentary
intranet. Use \code{fetch_crb_traffic_intranet} to retrieve equivalent data
for the Parliamentary intranet.
}
