% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsite.R
\name{fetch_ms_traffic}
\alias{fetch_ms_traffic}
\title{Download traffic data for all pages in the Commons Library microsite}
\usage{
fetch_ms_traffic(start_date, end_date, internal = FALSE,
  by_date = FALSE, by_page = FALSE, merge_paths = FALSE)
}
\arguments{
\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}

\item{internal}{A boolean indicating whether to return only the results for
traffic from internal parliamentary networks. The default is FALSE.}

\item{by_date}{A boolean indicating whether to return the results broken
down by date. The default is FALSE.}

\item{by_page}{A boolean indicating whether to return the results broken
down by individual page. The default is FALSE.}

\item{merge_paths}{A boolean indicating whether to aggregate metrics for all
pages that have the same root path i.e. for all pages whose paths differ
only by their query string. This parameter is ignored if \code{by_page} is
set to FALSE. Note that while merging paths is necessary for analysis of
individual pages it can introduce small errors in the number of users by
page, as the same user may visit the same page through URLs with different
query strings. The default value is FALSE.}
}
\value{
A tibble of traffic metrics.
}
\description{
\code{fetch_ms_traffic} downloads data on traffic metrics for all posts on
the Commons Library microsite during the given dates and returns the data
as a tibble.
}
