% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/briefings.R
\name{fetch_rb_traffic_intranet_by_type}
\alias{fetch_rb_traffic_intranet_by_type}
\title{Download traffic data for different types of Commons research briefings on
the Parliamentary intranet}
\usage{
fetch_rb_traffic_intranet_by_type(type_regexp, start_date, end_date,
  detailed = FALSE)
}
\arguments{
\item{type_regexp}{That regular expression that describes the page path for
one or more briefing types.}

\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}

\item{detailed}{A boolean indicating whether to return the results by date
in total, or broken down by date and page. The default is FALSE.}
}
\value{
A tibble of traffic metrics.
}
\description{
\code{fetch_rb_traffic_intranet_by_type} downloads data on traffic metrics
for specific types of research breifing based on their id prefixes during
the given dates and returns the data as a tibble. The specific types of
briefings to return are defined with a regular expression that identifies
their id prefixes within the page path.
}
\details{
Note that this function only returns data for the Parliamentary intranet.
Use \code{fetch_rb_traffic_public_by_type} to retrieve equivalent data
for the main Parliamentary website.
}
