% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic.R
\name{fetch_traffic}
\alias{fetch_traffic}
\title{Download traffic data for a given view}
\usage{
fetch_traffic(view_id, start_date, end_date, dimensions = c("date"),
  dim_filters = NULL)
}
\arguments{
\item{view_id}{The view id in Google Analytics}

\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}

\item{dimensions}{A vector of dimensions to return. The default is "date".}

\item{dim_filters}{A set of dimension filters to constrain the results. The
default is NULL.}
}
\value{
A tibble of traffic metrics.
}
\description{
\code{fetch_traffic} downloads data on traffic metrics for a given view
during the given dates and returns the data as a tibble.
}
\keyword{internal}
