% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsite-traffic.R
\name{fetch_traffic_for_ms}
\alias{fetch_traffic_for_ms}
\title{Download traffic data for a post with the given URL on the Commons Library
microsite}
\usage{
fetch_traffic_for_ms(
  url,
  start_date,
  end_date,
  internal = FALSE,
  by_date = FALSE,
  anti_sample = FALSE,
  use_resource_quotas = FALSE
)
}
\arguments{
\item{url}{The URL of a page for which traffic data is requested.}

\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}

\item{internal}{A boolean indicating whether to return only the results for
traffic from internal parliamentary networks. The default is FALSE.}

\item{by_date}{A boolean indicating whether to return the results broken
down by date. The default is FALSE.}

\item{anti_sample}{A boolean indicating whether to use googleAnalyticsR's
anti-sample feature, which chunks API calls to keep the number of records
requested under the API limits that trigger sampling. This makes the
download process slower but ensures that all records are returned. Only
use this feature if you see that an API request triggers sampling without
it. The default is FALSE.}

\item{use_resource_quotas}{A boolean indicating whether to use the resource
quotas in Parliament's Google Analytics account to prevent sampling.
This is a faster and more effective way to disable sampling than using
\code{anti_sample}, but using resource quotas consumes tokens from a
limited daily quota. Use this when \code{anti_sample} still fails to
prevent sampling or is taking too long. Note that using resource quotas
takes precendence over anti-samping: if \code{use_resource_quotas} is TRUE
\code{anti_sample} is automatically set to FALSE. The default is FALSE.}
}
\value{
A tibble of traffic metrics.
}
\description{
\code{fetch_traffic_for_ms} downloads data on traffic metrics for a given
post url on the Commons Library microsite during the given dates and returns
the data as a tibble.
}
