% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsite.R
\name{fetch_traffic_for_ms}
\alias{fetch_traffic_for_ms}
\title{Download traffic data for a post with the given URL on the Commons Library
microsite}
\usage{
fetch_traffic_for_ms(url, start_date, end_date, internal = FALSE,
  by_date = FALSE)
}
\arguments{
\item{url}{The URL of a page for which traffic data is requested.}

\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}

\item{internal}{A boolean indicating whether to return only the results for
traffic from internal parliamentary networks. The default is FALSE.}

\item{by_date}{A boolean indicating whether to return the results broken
down by date. The default is FALSE.}
}
\value{
A tibble of traffic metrics.
}
\description{
\code{fetch_traffic_for_ms} downloads data on traffic metrics for a given
post url on the Commons Library microsite during the given dates and returns
the data as a tibble.
}
