% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/briefings.R
\name{fetch_traffic_for_rb}
\alias{fetch_traffic_for_rb}
\title{Download traffic data for a research briefing with the given URL on the main
Parliament website and the Parliamentary intranet}
\usage{
fetch_traffic_for_rb(url, start_date, end_date, internal = FALSE,
  by_date = FALSE, combined = FALSE, anti_sample = FALSE)
}
\arguments{
\item{url}{The URL of a page for which traffic data is requested.}

\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}

\item{internal}{A boolean indicating whether to return only the results for
traffic from internal parliamentary networks. The default is FALSE.}

\item{by_date}{A boolean indicating whether to return the results broken
down by date. The default is FALSE.}

\item{combined}{A boolean indicating whether to combine the totals from
the website and the intranet or to report them separately. Note that
combining the traffic across both properties can introduce errors in
the number of users, as the same user may visit pages on both properties.
The default is FALSE.}

\item{anti_sample}{A boolean indicating whether to use googleAnalyticsR's
anti-sample feature, which chunks API calls to keep the number of records
requested under the API limits that trigger sampling. This makes the
download process slower but ensures that all records are returned. Only
use this feature if you see that an API request triggers sampling without
it. The default is FALSE.}
}
\value{
A tibble of traffic metrics.
}
\description{
\code{fetch_traffic_for_rb} downloads data on traffic metrics for a given
research briefing url on both the main Parliament website and the
Parliamentary intranet during the given dates and returns the data as a
tibble.
}
\details{
The data can either be combined so that each result appears once with
totals across both the Parliament website and the intranet, or reported
separately with separate rows for the website and the intranet.
}
