% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/briefings.R
\name{fetch_traffic_for_rb}
\alias{fetch_traffic_for_rb}
\title{Download traffic data for a research briefing with the given URL on the main
Parliament website and the Parliamentary intranet}
\usage{
fetch_traffic_for_rb(url, start_date, end_date, internal = FALSE,
  by_date = FALSE, combine = FALSE, anti_sample = FALSE)
}
\arguments{
\item{url}{The URL of a page for which traffic data is requested.}

\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}

\item{internal}{A boolean indicating whether to return only the results for
traffic from internal parliamentary networks. The default is FALSE.}

\item{by_date}{A boolean indicating whether to return the results broken
down by date. The default is FALSE.}

\item{combine}{A boolean indicating whether to combine the totals from
different properties or to report them separately. Note that combining the
traffic across properties can introduce errors in the number of users, as
the same user may visit pages on different properties. The default is
FALSE.}

\item{anti_sample}{A boolean indicating whether to use googleAnalyticsR's
anti-sample feature, which chunks API calls to keep the number of records
requested under the API limits that trigger sampling. This makes the
download process slower but ensures that all records are returned. Only
use this feature if you see that an API request triggers sampling without
it. The default is FALSE.}
}
\value{
A tibble of traffic metrics.
}
\description{
\code{fetch_traffic_for_rb} downloads data on traffic metrics for a given
research briefing url on both the main Parliament website and the
Parliamentary intranet during the given dates and returns the data as a
tibble.
}
\details{
By default, traffic figures are reported separately for each website
property in Google Analytics that contains some of the requested data. You
can use the \code{combine} argument to optionally combine traffic figures
so that each result appears only once with figures totalled across all
relevant properties. This makes it easier to produce aggregate figures for
traffic across a number of properties, but may introduce errors in the
number of users, as the same user may visit pages on different properties.
}
