% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/briefings.R
\name{fetch_traffic_for_rb_intranet}
\alias{fetch_traffic_for_rb_intranet}
\title{Download traffic data for a research briefing with the given URL on the
Parliamentary intranet}
\usage{
fetch_traffic_for_rb_intranet(url, start_date, end_date)
}
\arguments{
\item{url}{The URL of a page for which traffic data is requested.}

\item{start_date}{The start date as an ISO 8601 string.}

\item{end_date}{The end date as an ISO 8601 string.}
}
\value{
A tibble of traffic metrics.
}
\description{
\code{fetch_traffic_for_rb_intranet} downloads data on traffic metrics for a
given research briefing url on the main Parliament website during the
given dates and returns the data as a tibble.
}
